style-property-aware hit testing, fix #316
map.featuresAt now includes features whose rendered representation
matches the query, not features whose geometry matches the query.

A query with `radius: 0` will now match lines and circles if the point
is within the rendered line.

also fix #2053
It now checks intersection based on the render type not the geometry
type. A polygon that is rendered as a line will only match if the query
matches the line. It will not include it if the query only matches in
the internal part of the polygon.

implemented:
circle-radius
circle-translate
fill-translate
line-width
line-offset
line-translate

not implemented yet (hard with the current symbol index):
text-translate
icon-translate